+++ Source/Views/Knobs/LabeledKnob.cpp
@@ -20,7 +20,7 @@ void LabeledKnob::paint(juce::Graphics &/*g*/) {}
 
 void LabeledKnob::resized() {
     knob.setTextBoxStyle(juce::Slider::TextEntryBoxPosition::TextBoxBelow, true,
-                         getWidth() * .5, getHeight() * labelHeight);
+                         (int)(getWidth() * .5), (int)(getHeight() * labelHeight));
     label.setFont(juce::Font(juce::Font::getDefaultMonospacedFontName(),
                              getHeight() * labelHeight * 0.7f,
                              juce::Font::plain));
